enterprise/lifecycle: rename frontend reviews module to lifecycle
